// USER_SRV/src/constants/enums.ts
export enum StatusMessage {
    invalid = "is/are empty",
    notFound = "not found",
    unknowun = "Something went wrong",
    success = "with success",
    missmatch = "missmatch",
    duplicate = "already exists",
    expired = "has expired"
}

export enum Status {
    OK = "OK",
    NOTOK = "NOTOK",
}